modif cast

+++ nibblercore/src/Nibbler.cpp
@@ -32,7 +32,7 @@ Nibbler::Nibbler(const std::vector<std::string>& av, const std::vector<std::stri
     _win.y() = _win.x() / (gamesize.x() / gamesize.y());
   else if (gamesize.y() / gamesize.x() != 0)
     _win.x() = _win.y() / (gamesize.y() / gamesize.x());
-  _time.setFps((((float)gamesize.x()) + ((float)gamesize.y())) * ((float)15) / (((float)50) + ((float)(50))));
+  _time.setFps((gamesize.x() + gamesize.y()) * 15.0f / (50.0f + 50.0f));
   _lib = new DLLoader<IGui*>(lib);
 }
 
@@ -44,7 +44,7 @@ void Nibbler::run()
   bool acted;
   float add;
 
-  add = ((((float)_gamesize.x()) + ((float)_gamesize.y())) * ((float)4) / (((float)50) + ((float)(50))));
+  add = ((_gamesize.x() + _gamesize.y()) * 4.0f / (50.0f + 50.0f));
 
   if ((gui = _lib->getInstance()) == NULL)
     throw nFault("Could not create the library windows", true);
